Optimize union operator for empty operands
Fixes #264
